refactor how venn diagram is used in cohort comparison
