+++ packages/portal-proto/src/features/cohortComparison/CohortCard.tsx
@@ -1,12 +1,8 @@
-import dynamic from "next/dynamic";
-import { LoadingOverlay, Paper } from "@mantine/core";
+import { Paper } from "@mantine/core";
 import { FIELD_LABELS } from "src/fields";
 import CohortVennDiagram from "./CohortVennDiagram";
 import Link from "next/link";
 import { CohortComparisonType } from "./CohortComparison";
-const VennDiagram = dynamic(() => import("@/features/charts/VennDiagram"), {
-  ssr: false,
-});
 
 interface CohortCardProps {
   readonly selectedCards: Record<string, boolean>;
@@ -18,14 +14,6 @@ interface CohortCardProps {
   readonly caseSetIds: string[];
   readonly casesFetching: boolean;
 }
-
-const DEFAULT_CHART_DATA = [
-  { key: "S1_minus_S2", value: "...", highlighted: false },
-  { key: "S2_minus_S1", value: "...", highlighted: false },
-  { key: "S1_intersect_S2", value: "...", highlighted: false },
-];
-
-const LABELS = ["S₁", "S₂"];
 
 const CohortCard: React.FC<CohortCardProps> = ({
   selectedCards,
@@ -89,22 +77,11 @@ const CohortCard: React.FC<CohortCardProps> = ({
       </div>
       <hr />
 
-      <div className="relative">
-        <LoadingOverlay
-          visible={casesFetching || counts.length === 0}
-          zIndex={1}
-        />
-        {caseSetIds.length > 0 && !(casesFetching || counts.length === 0) ? (
-          <CohortVennDiagram caseSetIds={caseSetIds} cohorts={cohorts} />
-        ) : (
-          <VennDiagram
-            chartData={DEFAULT_CHART_DATA}
-            labels={LABELS}
-            ariaLabel="The Venn diagram displays the number of cases shared between the cohorts."
-            interactable={false}
-          />
-        )}
-      </div>
+      <CohortVennDiagram
+        caseSetIds={caseSetIds}
+        cohorts={cohorts}
+        isLoading={casesFetching || counts.length === 0}
+      />
 
       <div className="-mt-8 mb-2 z-10 flex justify-center relative">
         <Link
