+++ packages/portal-proto/src/features/facets/NumericRangeFacet/FromTo.tsx
@@ -212,7 +212,7 @@ const FromTo: React.FC<FromToProps> = ({
           (fromValueInDays !== undefined && fromValueInDays >= WARNING_DAYS),
       );
     }
-  }, [field, form.values.toValue, form.values.fromValue, units]);
+  }, [field, form.values.toValue, form.values.fromValue, units, queryInYears]);
 
   /**
    * Handle Apply button which will add/update/remove a range filter to the field.
